Refactor Synchronizer out of MediaUtil
